+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Harden sources based on more fuzz testing ([#1774](../../pull/1774))
 - Default to not caching source in notebooks ([#1776](../../pull/1776))
 - Automatically set the JUPYTER_PROXY value ([#1781](../../pull/1781))
+- Add a general channelNames property to tile sources ([#1783](../../pull/1783))
 
 ### Changes